@@ -13,9 +13,9 @@ import Foundation
 /// A class that can accept invocations and forward to Web3
 public protocol SolidityFunctionHandler: AnyObject {
     var address: EthereumAddress? { get }
-    func call(_ call: EthereumCall, outputs: [SolidityFunctionParameter], block: EthereumQuantityTag, completion: @escaping ([String: Any]?, Error?) -> Void)
-    func send(_ transaction: EthereumTransaction, completion: @escaping (EthereumData?, Error?) -> Void)
-    func estimateGas(_ call: EthereumCall, completion: @escaping (EthereumQuantity?, Error?) -> Void)
+    func call(_ call: EthereumCall, outputs: [SolidityFunctionParameter], block: EthereumQuantityTag) async -> ([String: Any]?, Error?)
+    func send(_ transaction: EthereumTransaction) async -> (EthereumData?, Error?)
+    func estimateGas(_ call: EthereumCall) async -> (EthereumQuantity?, Error?)
 }
 
 public protocol SolidityParameter {
