@@ -10,10 +10,6 @@ import Quick
 import Nimble
 @testable import Web3
 import BigInt
-import PromiseKit
-#if canImport(Web3PromiseKit)
-    @testable import Web3PromiseKit
-#endif
 
 class Web3HttpTests: QuickSpec {
 
@@ -27,7 +23,8 @@ class Web3HttpTests: QuickSpec {
             context("web3 client version") {
 
                 waitUntil(timeout: 2.0) { done in
-                    web3.clientVersion { response in
+                    Task {
+                        let response = await web3.clientVersion()
                         it("should be status success") {
                             expect(response.status.isSuccess) == true
                         }
@@ -44,7 +41,8 @@ class Web3HttpTests: QuickSpec {
             context("net version") {
 
                 waitUntil(timeout: 2.0) { done in
-                    web3.net.version { response in
+                    Task {
+                        let response = await web3.net.version()
                         it("should be status ok") {
                             expect(response.status.isSuccess) == true
                         }
@@ -64,7 +62,8 @@ class Web3HttpTests: QuickSpec {
             context("net peer count") {
 
                 waitUntil(timeout: 2.0) { done in
-                    web3.net.peerCount { response in
+                    Task {
+                        let response = await web3.net.peerCount()
                         it("should be status ok") {
                             expect(response.status.isSuccess) == true
                         }
@@ -84,7 +83,9 @@ class Web3HttpTests: QuickSpec {
             context("eth protocol version") {
 
                 waitUntil(timeout: 2.0) { done in
-                    web3.eth.protocolVersion { response in
+                    Task {
+                        let response = await web3.eth.protocolVersion()
+                        
                         it("should be status ok") {
                             expect(response.status.isSuccess) == true
                         }
@@ -101,7 +102,8 @@ class Web3HttpTests: QuickSpec {
             context("eth syncing") {
 
                 waitUntil(timeout: 2.0) { done in
-                    web3.eth.syncing { response in
+                    Task {
+                        let response = await web3.eth.syncing()
                         it("should be status ok") {
                             expect(response.status.isSuccess) == true
                         }
@@ -131,7 +133,9 @@ class Web3HttpTests: QuickSpec {
             context("eth mining") {
 
                 waitUntil(timeout: 2.0) { done in
-                    web3.eth.mining { response in
+                    Task {
+                        let response = await web3.eth.mining()
+                        
                         it("should be status ok") {
                             expect(response.status.isSuccess) == true
                         }
@@ -152,7 +156,8 @@ class Web3HttpTests: QuickSpec {
             context("eth hashrate") {
 
                 waitUntil(timeout: 2.0) { done in
-                    web3.eth.hashrate { response in
+                    Task {
+                        let response = await web3.eth.hashrate()
                         it("should be status ok") {
                             expect(response.status.isSuccess) == true
                         }
@@ -173,7 +178,8 @@ class Web3HttpTests: QuickSpec {
             context("eth gas price") {
 
                 waitUntil(timeout: 2.0) { done in
-                    web3.eth.gasPrice { response in
+                    Task {
+                        let response = await web3.eth.gasPrice()
                         it("should be status ok") {
                             expect(response.status.isSuccess) == true
                         }
@@ -194,7 +200,8 @@ class Web3HttpTests: QuickSpec {
             context("eth accounts") {
 
                 waitUntil(timeout: 2.0) { done in
-                    web3.eth.accounts { response in
+                    Task {
+                        let response = await web3.eth.accounts()
                         it("should be status ok") {
                             expect(response.status.isSuccess) == true
                         }
@@ -215,7 +222,8 @@ class Web3HttpTests: QuickSpec {
             context("eth block number") {
 
                 waitUntil(timeout: 2.0) { done in
-                    web3.eth.blockNumber { response in
+                    Task {
+                        let response = await web3.eth.blockNumber()
                         it("should be status ok") {
                             expect(response.status.isSuccess) == true
                         }
@@ -243,7 +251,8 @@ class Web3HttpTests: QuickSpec {
                 }
 
                 waitUntil(timeout: 2.0) { done in
-                    web3.eth.getBalance(address: ethereumAddress, block: .block(4000000)) { response in
+                    Task {
+                        let response = await web3.eth.getBalance(address: ethereumAddress, block: .block(4000000))
                         it("should be status ok") {
                             expect(response.status.isSuccess) == true
                         }
@@ -256,6 +265,7 @@ class Web3HttpTests: QuickSpec {
 
                         // Tests done
                         done()
+                        
                     }
                 }
             }
@@ -271,7 +281,8 @@ class Web3HttpTests: QuickSpec {
                 }
 
                 waitUntil(timeout: 2.0) { done in
-                    web3.eth.getStorageAt(address: ethereumAddress, position: 0, block: .latest) { response in
+                    Task {
+                        let response = await web3.eth.getStorageAt(address: ethereumAddress, position: 0, block: .latest)
                         it("should be status ok") {
                             expect(response.status.isSuccess) == true
                         }
@@ -299,7 +310,8 @@ class Web3HttpTests: QuickSpec {
                 }
 
                 waitUntil(timeout: 2.0) { done in
-                    web3.eth.getTransactionCount(address: ethereumAddress, block: .block(4000000)) { response in
+                    Task {
+                        let response = await web3.eth.getTransactionCount(address: ethereumAddress, block: .block(4000000))
                         it("should be status ok") {
                             expect(response.status.isSuccess) == true
                         }
@@ -318,8 +330,9 @@ class Web3HttpTests: QuickSpec {
 
             context("eth get transaction count by hash") {
                 waitUntil(timeout: 2.0) { done in
-                    do {
-                        try web3.eth.getBlockTransactionCountByHash(blockHash: .string("0x596f2d863a893392c55b72b5ba29e9ba67bdaa13c31765f9119e850a62565960")) { response in
+                    Task {
+                        do {
+                            let response = try await web3.eth.getBlockTransactionCountByHash(blockHash: .string("0x596f2d863a893392c55b72b5ba29e9ba67bdaa13c31765f9119e850a62565960"))
                             it("should be status ok") {
                                 expect(response.status.isSuccess) == true
                             }
@@ -332,28 +345,30 @@ class Web3HttpTests: QuickSpec {
 
                             // Tests done
                             done()
+                        } catch {
+                            it("should not throw an error") {
+                                expect(false) == true
+                            }
+                            done()
                         }
-                    } catch {
-                        it("should not throw an error") {
-                            expect(false) == true
-                        }
-                        done()
                     }
                 }
             }
 
             context("eth get transaction count by number") {
                 waitUntil(timeout: 2.0) { done in
-                    firstly {
-                        web3.eth.getBlockTransactionCountByNumber(block: .block(5397389))
-                    }.done { count in
-                        it("should be count 88") {
-                            expect(count) == 88
-                        }
-                        done()
-                    }.catch { error in
-                        it("should not fail") {
-                            expect(false) == true
+                    Task {
+                        let response = await web3.eth.getBlockTransactionCountByNumber(block: .block(5397389))
+                        
+                        switch response.status {
+                        case .success(let count):
+                            it("should be count 88") {
+                                expect(count) == 88
+                            }
+                        case .failure(let error):
+                            it("should not fail") {
+                                fail(error.localizedDescription)
+                            }
                         }
                         done()
                     }
@@ -362,16 +377,17 @@ class Web3HttpTests: QuickSpec {
 
             context("eth get uncle count by block hash") {
                 waitUntil(timeout: 2.0) { done in
-                    firstly {
-                        try web3.eth.getUncleCountByBlockHash(blockHash: .string("0xd8cdd624c5b4c5323f0cb8536ca31de046e3e4a798a07337489bab1bb3d822f0"))
-                    }.done { count in
-                        it("should include one uncle") {
-                            expect(count) == 1
-                        }
-                        done()
-                    }.catch { error in
-                        it("should not fail") {
-                            expect(false) == true
+                    Task {
+                        let response = try await web3.eth.getUncleCountByBlockHash(blockHash: .string("0xd8cdd624c5b4c5323f0cb8536ca31de046e3e4a798a07337489bab1bb3d822f0"))
+                        switch response.status {
+                        case .success(let count):
+                            it("should include one uncle") {
+                                expect(count) == 1
+                            }
+                        case .failure(let error):
+                            it("should not fail") {
+                                fail(error.localizedDescription)
+                            }
                         }
                         done()
                     }
@@ -380,16 +396,18 @@ class Web3HttpTests: QuickSpec {
 
             context("eth get uncle count by block number") {
                 waitUntil(timeout: 2.0) { done in
-                    firstly {
-                        web3.eth.getUncleCountByBlockNumber(block: .block(5397429))
-                    }.done { count in
-                        it("should include one uncle") {
-                            expect(count) == 1
-                        }
-                        done()
-                    }.catch { error in
-                        it("should not fail") {
-                            expect(false) == true
+                    Task {
+                        let response = await web3.eth.getUncleCountByBlockNumber(block: .block(5397429))
+                        
+                        switch response.status {
+                        case .success(let count):
+                            it("should include one uncle") {
+                                expect(count) == 1
+                            }
+                        case .failure(let error):
+                            it("should not fail") {
+                                fail(error.localizedDescription)
+                            }
                         }
                         done()
                     }
@@ -398,17 +416,18 @@ class Web3HttpTests: QuickSpec {
 
             context("eth get code") {
                 waitUntil(timeout: 2.0) { done in
-                    firstly {
-                        try web3.eth.getCode(address: EthereumAddress(hex: "0x2e704bF506b96adaC7aD1df0db461344146a4657", eip55: true), block: .block(5397525))
-                    }.done { code in
-                        it("should be the expected data") {
-                            let data: EthereumData? = try? .string("0x60606040526004361061006c5763ffffffff7c0100000000000000000000000000000000000000000000000000000000600035041663022914a78114610071578063173825d9146100a457806341c0e1b5146100c55780637065cb48146100d8578063aa1e84de146100f7575b600080fd5b341561007c57600080fd5b610090600160a060020a036004351661015a565b604051901515815260200160405180910390f35b34156100af57600080fd5b6100c3600160a060020a036004351661016f565b005b34156100d057600080fd5b6100c36101b7565b34156100e357600080fd5b6100c3600160a060020a03600435166101ea565b341561010257600080fd5b61014860046024813581810190830135806020601f8201819004810201604051908101604052818152929190602084018383808284375094965061023595505050505050565b60405190815260200160405180910390f35b60006020819052908152604090205460ff1681565b600160a060020a03331660009081526020819052604090205460ff16151561019657600080fd5b600160a060020a03166000908152602081905260409020805460ff19169055565b600160a060020a03331660009081526020819052604090205460ff1615156101de57600080fd5b33600160a060020a0316ff5b600160a060020a03331660009081526020819052604090205460ff16151561021157600080fd5b600160a060020a03166000908152602081905260409020805460ff19166001179055565b6000816040518082805190602001908083835b602083106102675780518252601f199092019160209182019101610248565b6001836020036101000a0380198251168184511617909252505050919091019250604091505051809103902090509190505600a165627a7a7230582085affe2ee33a8eb3900e773ef5a0d7f1bc95448e61a845ef36e00e6d6b4872cf0029")
-                            expect(code) == data
-                        }
-                        done()
-                    }.catch { error in
-                        it("should not fail") {
-                            expect(false) == true
+                    Task {
+                        let response = try await web3.eth.getCode(address: EthereumAddress(hex: "0x2e704bF506b96adaC7aD1df0db461344146a4657", eip55: true), block: .block(5397525))
+                        switch response.status {
+                        case .success(let code):
+                            it("should be the expected data") {
+                                let data: EthereumData? = try? .string("0x60606040526004361061006c5763ffffffff7c0100000000000000000000000000000000000000000000000000000000600035041663022914a78114610071578063173825d9146100a457806341c0e1b5146100c55780637065cb48146100d8578063aa1e84de146100f7575b600080fd5b341561007c57600080fd5b610090600160a060020a036004351661015a565b604051901515815260200160405180910390f35b34156100af57600080fd5b6100c3600160a060020a036004351661016f565b005b34156100d057600080fd5b6100c36101b7565b34156100e357600080fd5b6100c3600160a060020a03600435166101ea565b341561010257600080fd5b61014860046024813581810190830135806020601f8201819004810201604051908101604052818152929190602084018383808284375094965061023595505050505050565b60405190815260200160405180910390f35b60006020819052908152604090205460ff1681565b600160a060020a03331660009081526020819052604090205460ff16151561019657600080fd5b600160a060020a03166000908152602081905260409020805460ff19169055565b600160a060020a03331660009081526020819052604090205460ff1615156101de57600080fd5b33600160a060020a0316ff5b600160a060020a03331660009081526020819052604090205460ff16151561021157600080fd5b600160a060020a03166000908152602081905260409020805460ff19166001179055565b6000816040518082805190602001908083835b602083106102675780518252601f199092019160209182019101610248565b6001836020036101000a0380198251168184511617909252505050919091019250604091505051809103902090509190505600a165627a7a7230582085affe2ee33a8eb3900e773ef5a0d7f1bc95448e61a845ef36e00e6d6b4872cf0029")
+                                expect(code) == data
+                            }
+                        case .failure(let error):
+                            it("should not fail") {
+                                fail(error.localizedDescription)
+                            }
                         }
                         done()
                     }
@@ -417,8 +436,8 @@ class Web3HttpTests: QuickSpec {
 
             context("eth call") {
                 waitUntil(timeout: 2.0) { done in
-                    firstly {
-                        try EthereumCall(
+                    Task {
+                        let call = try EthereumCall(
                             from: nil,
                             to: EthereumAddress(hex: "0x2e704bf506b96adac7ad1df0db461344146a4657", eip55: false),
                             gas: nil,
@@ -427,18 +446,19 @@ class Web3HttpTests: QuickSpec {
                             data: EthereumData(
                                 ethereumValue: "0xaa1e84de000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000046461766500000000000000000000000000000000000000000000000000000000"
                             )
-                        ).promise
-                    }.then { call in
-                        web3.eth.call(call: call, block: .latest)
-                    }.done { data in
-                        let expectedData: EthereumData = try .string("0x5e2393c41c2785095aa424cf3e033319468b6dcebda65e61606ee2ae2a198a87")
-                        it("should be the expected data") {
-                            expect(data) == expectedData
-                        }
-                        done()
-                    }.catch { error in
-                        it("should not fail") {
-                            expect(false) == true
+                        )
+                        let response = await web3.eth.call(call: call, block: .latest)
+                        
+                        switch response.status {
+                        case .success(let data):
+                            let expectedData: EthereumData = try .string("0x5e2393c41c2785095aa424cf3e033319468b6dcebda65e61606ee2ae2a198a87")
+                            it("should be the expected data") {
+                                expect(data) == expectedData
+                            }
+                        case .failure(let error):
+                            it("should not fail") {
+                                fail(error.localizedDescription)
+                            }
                         }
                         done()
                     }
@@ -447,8 +467,8 @@ class Web3HttpTests: QuickSpec {
 
             context("eth estimate gas") {
                 waitUntil(timeout: 2.0) { done in
-                    firstly {
-                        try EthereumCall(
+                    Task {
+                        let call = try EthereumCall(
                             from: nil,
                             to: EthereumAddress(hex: "0x2e704bf506b96adac7ad1df0db461344146a4657", eip55: false),
                             gas: nil,
@@ -457,18 +477,18 @@ class Web3HttpTests: QuickSpec {
                             data: EthereumData(
                                 ethereumValue: "0xaa1e84de000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000046461766500000000000000000000000000000000000000000000000000000000"
                             )
-                        ).promise
-                    }.then { call in
-                        web3.eth.estimateGas(call: call)
-                    }.done { quantity in
-                        let expectedQuantity: EthereumQuantity = try .string("0x56d4")
-                        it("should be the expected quantity") {
-                            expect(quantity) == expectedQuantity
-                        }
-                        done()
-                    }.catch { error in
-                        it("should not fail") {
-                            expect(false) == true
+                        )
+                        let response = await web3.eth.estimateGas(call: call)
+                        switch response.status {
+                        case .success(let quantity):
+                            let expectedQuantity: EthereumQuantity = try .string("0x56d4")
+                            it("should be the expected quantity") {
+                                expect(quantity) == expectedQuantity
+                            }
+                        case .failure(let error):
+                            it("should not fail") {
+                                fail(error.localizedDescription)
+                            }
                         }
                         done()
                     }
